requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.0
accelerate>=0.28.0
datasets>=2.18.0
pillow>=10.0.0

# Config
pyyaml>=6.0
hydra-core>=1.3.0

# MLOps
mlflow>=2.10.0
wandb>=0.16.0
dvc>=3.0.0

# Serving
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gradio>=4.0.0

# Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.4.0

# Utils
tqdm>=4.66.0
numpy>=1.24.0
